Switch to `aiodocker`
This slightly change the error handling because if the command return with a non-zero exit code this does not raise a `DockerError` exception anymore (for example when the snippet has a syntax error), but will now be reported as a `ManimError`. Old behaviour can be found back with a few added lines if needed.

Close https://github.com/ManimCommunity/DiscordManimator/issues/41