Support async docker communication (#43)
* Update dependencies

aiohttp (3.8.1) was not supporting PEP 517 builds. (on M1)

* Switch to `aiodocker`

This slightly change the error handling because if the command return with a non-zero exit code this does not raise a `DockerError` exception anymore (for example when the snippet has a syntax error), but will now be reported as a `ManimError`. Old behaviour can be found back with a few added lines if needed.

Close https://github.com/ManimCommunity/DiscordManimator/issues/41

* Update and format the repository with black

Update was necessary because of a bug with the previously in use black version: https://stackoverflow.com/a/71674345/11955835

* Update cogs/render_codeblock.py

---------

Co-authored-by: kraktus <kraktus@users.noreply.github.com>
Co-authored-by: Benjamin Hackl <devel@benjamin-hackl.at>